Refactor string_manipulation library (2/2)
score/memory contained many libraries that are normally used for string manipulation and not for memory operations.

Moves these libraries to memory/string_manipulation.

This breaks public API. Users will have to adapt.
