+++ score/json/internal/model/BUILD
@@ -36,8 +36,8 @@ cc_library(
     visibility = ["@score_baselibs//score/json:__subpackages__"],
     deps = [
         "@score_baselibs//score/language/futurecpp",
-        "@score_baselibs//score/memory:string_comparison_adaptor",
         "@score_baselibs//score/result",
+        "@score_baselibs//score/string_manipulation:string_comparison_adaptor",
     ],
 )
