+++ score/string_manipulation/BUILD
@@ -11,6 +11,9 @@
 # SPDX-License-Identifier: Apache-2.0
 # *******************************************************************************
 
+load("@rules_cc//cc:defs.bzl", "cc_library", "cc_test")
+load("@score_baselibs//:bazel/unit_tests.bzl", "cc_unit_test", "cc_unit_test_suites_for_host_and_qnx")
+load("@score_baselibs//score/language/safecpp:toolchain_features.bzl", "COMPILER_WARNING_FEATURES")
 load("@score_baselibs//score/quality/clang_tidy:extra_checks.bzl", "clang_tidy_extra_checks")
 
 clang_tidy_extra_checks(
@@ -23,4 +26,94 @@ clang_tidy_extra_checks(
         "spp_code_style_check_variable_names",
     ],
     tidy_config_file = ".clang-tidy-extra",
+)
+
+cc_library(
+    name = "string_literal",
+    hdrs = ["string_literal.h"],
+    tags = ["FFI"],
+    visibility = [
+        "//visibility:public",
+    ],
+)
+
+cc_library(
+    name = "string_comparison_adaptor",
+    srcs = [
+        "string_comparison_adaptor.cpp",
+    ],
+    hdrs = [
+        "string_comparison_adaptor.h",
+    ],
+    features = COMPILER_WARNING_FEATURES,
+    implementation_deps = ["@score_baselibs//score/language/futurecpp"],
+    tags = ["FFI"],
+    visibility = [
+        "//visibility:public",
+    ],
+    deps = [":string_literal"],
+)
+
+cc_test(
+    name = "string_comparison_adaptor_unit_test",
+    srcs = [
+        "string_comparison_adaptor_test.cpp",
+    ],
+    features = [
+        "aborts_upon_exception",
+        "treat_warnings_as_errors",
+        "strict_warnings",
+        "additional_warnings",
+    ],
+    tags = ["unit"],
+    deps = [
+        ":string_comparison_adaptor",
+        "@googletest//:gtest_main",
+        "@score_baselibs//score/language/futurecpp",
+    ],
+)
+
+cc_library(
+    name = "split_string_view",
+    srcs = [
+        "split_string_view.cpp",
+    ],
+    hdrs = [
+        "split_string_view.h",
+    ],
+    features = COMPILER_WARNING_FEATURES,
+    implementation_deps = ["@score_baselibs//score/language/futurecpp"],
+    tags = ["FFI"],
+    visibility = [
+        "@score_baselibs//score/mw/log:__subpackages__",
+    ],
+)
+
+cc_test(
+    name = "split_string_view_unit_test",
+    srcs = [
+        "split_string_view_test.cpp",
+    ],
+    features = [
+        "aborts_upon_exception",
+        "treat_warnings_as_errors",
+        "strict_warnings",
+        "additional_warnings",
+    ],
+    tags = ["unit"],
+    deps = [
+        ":split_string_view",
+        "@googletest//:gtest_main",
+    ],
+)
+
+cc_unit_test_suites_for_host_and_qnx(
+    name = "unit_test_suite",
+    cc_unit_tests = [
+        ":split_string_view_unit_test",
+        ":string_comparison_adaptor_unit_test",
+    ],
+    visibility = [
+        "//visibility:public",
+    ],
 )
